+++ packages/components/modules/messages/web/EditGroup/constants.ts
@@ -1,20 +1,15 @@
-import { z } from 'zod'
+import z from 'zod'
 
-import { TitleAndImage } from '../__shared__/types'
-import { EditGroupUpload } from './types'
-
-export const FORM_VALUE: Record<keyof EditGroupUpload, keyof EditGroupUpload> &
-  Record<keyof TitleAndImage, keyof TitleAndImage> = {
-  title: 'title',
-  addParticipants: 'addParticipants',
-  removeParticipants: 'removeParticipants',
-  image: 'image',
-}
+import {
+  CREATE_OR_EDIT_GROUP_FORM_VALUE,
+  DEFAULT_CREATE_OR_EDIT_GROUP_FORM_VALIDATION,
+} from '../__shared__/constants'
+import { CreateOrEditGroup } from '../__shared__/types'
 
 export const getDefaultFormValues = (
   title: string,
   image: string | undefined,
-): EditGroupUpload => ({
+): CreateOrEditGroup => ({
   title,
   addParticipants: [],
   removeParticipants: [],
@@ -22,8 +17,6 @@ export const getDefaultFormValues = (
 })
 
 export const DEFAULT_FORM_VALIDATION = z.object({
-  [FORM_VALUE.title]: z.string().min(1, { message: 'Please enter a title' }),
-  [FORM_VALUE.addParticipants]: z.array(z.any()),
-  [FORM_VALUE.removeParticipants]: z.array(z.any()),
-  [FORM_VALUE.image]: z.any(),
+  ...DEFAULT_CREATE_OR_EDIT_GROUP_FORM_VALIDATION.shape,
+  [CREATE_OR_EDIT_GROUP_FORM_VALUE.participants]: z.any(),
 })