+++ packages/components/modules/messages/web/EditGroup/index.tsx
@@ -1,41 +1,74 @@
 'use client'
 
-import { FC } from 'react'
+import { FC, useMemo, useState, useTransition } from 'react'
 
 import { useCurrentProfile } from '@baseapp-frontend/authentication'
 import { IconButton } from '@baseapp-frontend/design-system/components/web/buttons'
 import { CheckMarkIcon, CloseIcon } from '@baseapp-frontend/design-system/components/web/icons'
+import { useResponsive } from '@baseapp-frontend/design-system/hooks/web'
 import { filterDirtyValues, setFormRelayErrors, useNotification } from '@baseapp-frontend/utils'
 
 import { zodResolver } from '@hookform/resolvers/zod'
 import { Box, Typography } from '@mui/material'
 import { useForm } from 'react-hook-form'
-import { usePreloadedQuery } from 'react-relay'
+import { usePaginationFragment, usePreloadedQuery } from 'react-relay'
 
+import { ChatRoomParticipantsPaginationQuery } from '../../../../__generated__/ChatRoomParticipantsPaginationQuery.graphql'
 import { GroupDetailsQuery as GroupDetailsQueryType } from '../../../../__generated__/GroupDetailsQuery.graphql'
+import { MembersListFragment$key } from '../../../../__generated__/MembersListFragment.graphql'
+import { ProfileNode } from '../../../profiles/common'
 import {
   GroupDetailsQuery,
+  MembersListFragment,
   useGroupNameAndAvatar,
   useRoomListSubscription,
   useUpdateChatRoomMutation,
 } from '../../common'
 import EditGroupTitleAndImage from '../__shared__/EditGroupTitleAndImage'
-import { DEFAULT_FORM_VALIDATION, FORM_VALUE, getDefaultFormValues } from './constants'
+import DefaultGroupChatMembersList from '../__shared__/GroupChatMembersList'
+import { CREATE_OR_EDIT_GROUP_FORM_VALUE as FORM_VALUE } from '../__shared__/constants'
+import AddMembersDialog from './AddMembersDialog'
+import AddMembersMobile from './AddMembersMobile'
+import { DEFAULT_FORM_VALIDATION, getDefaultFormValues } from './constants'
 import { HeaderContainer } from './styled'
 import { EditGroupProps } from './types'
 
 const EditGroup: FC<EditGroupProps & { profileId: string }> = ({
   profileId,
+  allProfilesRef,
   queryRef,
   roomId,
+  GroupChatMembersList = DefaultGroupChatMembersList,
+  GroupChatMembersListProps = {},
   onCancellation,
   onRemovalFromGroup,
   onValidSubmission,
 }) => {
   const { sendToast } = useNotification()
+  const [open, setOpen] = useState(false)
+  const smDown = useResponsive('down', 'sm')
   const { chatRoom: group } = usePreloadedQuery<GroupDetailsQueryType>(GroupDetailsQuery, queryRef)
   const { avatar, title } = useGroupNameAndAvatar(group)
   useRoomListSubscription({ profileId, connections: [], onRemoval: onRemovalFromGroup })
+
+  const {
+    data: membersList,
+    loadNext,
+    isLoadingNext,
+    hasNext,
+    refetch,
+  } = usePaginationFragment<ChatRoomParticipantsPaginationQuery, MembersListFragment$key>(
+    MembersListFragment,
+    group,
+  )
+
+  const participants = useMemo(
+    () =>
+      membersList?.participants?.edges?.map(
+        (edge: any) => edge?.node?.profile && edge.node.profile,
+      ) as ProfileNode[],
+    [membersList],
+  )
 
   const formReturn = useForm({
     defaultValues: getDefaultFormValues(title || '', avatar),
@@ -70,6 +103,7 @@ const EditGroup: FC<EditGroupProps & { profileId: string }> = ({
       }
       delete dirtyValues.image
     }
+    delete dirtyValues.participants
 
     commit({
       variables: {
@@ -103,9 +137,39 @@ const EditGroup: FC<EditGroupProps & { profileId: string }> = ({
   }
 
   const isEditButtonDisabled = !isValid || !isDirty
+  const [isPending, startTransition] = useTransition()
+  const handleAddMemberSuccess = () => {
+    setOpen(false)
+    startTransition(() => {
+      refetch?.({})
+    })
+  }
+
+  if (smDown && open)
+    return (
+      <AddMembersMobile
+        allProfilesRef={allProfilesRef}
+        onClose={() => setOpen(false)}
+        handleSubmitSuccess={handleAddMemberSuccess}
+        profileId={profileId}
+        roomId={roomId}
+        isPending={isPending}
+        existingMembers={participants}
+      />
+    )
 
   return (
     <Box>
+      <AddMembersDialog
+        open={open}
+        allProfilesRef={allProfilesRef}
+        onClose={() => setOpen(false)}
+        handleSubmitSuccess={handleAddMemberSuccess}
+        profileId={profileId}
+        roomId={roomId}
+        isPending={isPending}
+        existingMembers={participants}
+      />
       <HeaderContainer>
         <IconButton onClick={onCancellation} aria-label="cancel editing group">
           <CloseIcon sx={{ fontSize: '24px' }} />
@@ -133,6 +197,22 @@ const EditGroup: FC<EditGroupProps & { profileId: string }> = ({
         setValue={setValue}
         trigger={trigger}
         watch={watch}
+      />
+      <GroupChatMembersList
+        FORM_VALUE={FORM_VALUE}
+        setValue={setValue}
+        watch={watch}
+        currentParticipants={participants}
+        refetch={refetch}
+        membersLoadNext={loadNext}
+        membersHasNext={hasNext}
+        membersIsLoadingNext={isLoadingNext}
+        MembersListProps={{
+          allowAddMember: true,
+          onAddMemberClick: () => setOpen(true),
+          ...(GroupChatMembersListProps?.MembersListProps ?? {}),
+        }}
+        {...GroupChatMembersListProps}
       />
     </Box>
   )