+++ packages/components/modules/messages/web/MessagesList/MessagesGroup/SystemMessage/styled.tsx
@@ -3,7 +3,7 @@ import { Typography, styled } from '@mui/material'
 export const SystemMessageTypography = styled(Typography)(({ theme }) => ({
   alignSelf: 'center',
   backgroundColor: theme.palette.grey[300],
-  borderRadius: theme.spacing(1),
+  borderRadius: theme.spacing(0.5),
   marginBottom: theme.spacing(1),
   marginTop: theme.spacing(2),
   padding: theme.spacing(0.5, 1),